third_party/blink/web_tests/platform/generic/fullscreen/non-ancestor-iframe-expected.txt

This tests that full screen mode will not exit when a non-ancestor iframe is loaded in the page. Click enter full screen to start test.
EVENT(webkitfullscreenchange)
EVENT(load)
EXPECTED (document.webkitIsFullScreen == 'true') OK
END OF TEST
